requirements.txt
Django==3.1.7
Markdown==3.3.4
asgiref==3.3.2
django-filter==2.4.0
djangorestframework==3.12.4
pip==21.0.1
psycopg2==2.8.6
pytz==2021.1
setuptools==54.2.0
sqlparse==0.4.1
